+++ packages/frontend/src/pages/Profile/PartiesOverviewPage/CompanyDetails.tsx
@@ -7,33 +7,92 @@ import {
   Section,
   SettingsItem,
 } from '@altinn/altinn-components';
-import { Buildings2Icon, HandshakeIcon, HeartFillIcon, HeartIcon, InboxIcon } from '@navikt/aksel-icons';
-import { useNotificationSettings } from '../useNotificationSettings';
+import { Buildings2Icon, HeartFillIcon, HeartIcon, InboxIcon } from '@navikt/aksel-icons';
+import { useQueryClient } from '@tanstack/react-query';
+import type { PartyFieldsFragment } from 'bff-types-generated';
+import { useState } from 'react';
+import { QUERY_KEYS } from '../../../constants/queryKeys';
+import type { NotificationAccountsType } from '../NotificationsPage/AccountSettings';
+import { NotificationSettingsModal } from '../NotificationsPage/NotificationSettingsModal';
+import { useNotificationSettingsForParty } from '../useNotificationSettingsForParty';
 import { NotificationSetting } from './NotificationSettings';
 
-export const CompanyDetails = ({ ...props }: AccountDetailsProps) => {
-  const { id: partyUuid, uniqueId, parentId, items } = props;
-  const parentAccount = items?.find((item) => item.id === parentId);
+export interface CompanyDetailsProps extends AccountListItemProps {
+  party?: PartyFieldsFragment;
+  parentAccount?: NotificationAccountsType;
+  userId?: string;
+  alertEmailAddress?: string;
+  alertPhoneNumber?: string;
+  contactEmailAddress?: string;
+  contactPhoneNumber?: string;
+  address?: string;
+}
 
-  const { notificationSettings, isLoading: isLoadingNotificaitonSettings } = useNotificationSettings(partyUuid);
+export const CompanyDetails = ({
+  uniqueId,
+  party,
+  parentAccount,
+  favourite,
+  onToggleFavourite,
+  icon,
+  isCurrentEndUser,
+  type,
+  name,
+  id,
+}: CompanyDetailsProps) => {
+  const [notificationParty, setNotificationParty] = useState<NotificationAccountsType | null>(null);
+  const queryClient = useQueryClient();
+  const { notificationSettingsForParty, isLoading: isLoadingNotificaitonSettings } =
+    useNotificationSettingsForParty(id);
 
+  if (!party) {
+    return null;
+  }
+
+  const onSave = () => {
+    queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.NOTIFICATIONSETTINGSFORPARTY] });
+  };
+
+  if (!party) {
+    return null;
+  }
   return (
     <Section color="company" padding={6} spacing={2}>
-      <AccountToolbar {...props} id={partyUuid} />
+      <AccountToolbar
+        type={type}
+        name={name}
+        id={id}
+        favourite={favourite || false}
+        onToggleFavourite={onToggleFavourite}
+        icon={icon}
+        isCurrentEndUser={isCurrentEndUser || false}
+      />
       <Divider />
-      {!isLoadingNotificaitonSettings && notificationSettings ? (
-        <NotificationSetting notificationSetting={notificationSettings} />
+      {!isLoadingNotificaitonSettings && notificationSettingsForParty ? (
+        <NotificationSetting
+          notificationSetting={notificationSettingsForParty}
+          setNotificationParty={() => setNotificationParty(party as NotificationAccountsType)}
+        />
       ) : (
         <NotificationSetting
-          key={partyUuid}
+          key={id}
           notificationSetting={{
-            partyUuid: partyUuid,
+            partyUuid: id,
             emailAddress: '',
             phoneNumber: '',
           }}
+          setNotificationParty={() => setNotificationParty(party as NotificationAccountsType)}
         />
       )}
-
+      {notificationParty && (
+        <NotificationSettingsModal
+          notificationParty={
+            { ...notificationParty, notificationSettings: notificationSettingsForParty } as NotificationAccountsType
+          }
+          setNotificationParty={(updatedParty: NotificationAccountsType | null) => setNotificationParty(updatedParty)}
+          onSave={onSave}
+        />
+      )}
       <List size="sm">
         <Divider as="li" />
         <SettingsItem
@@ -53,18 +112,19 @@ export const CompanyDetails = ({ ...props }: AccountDetailsProps) => {
             />
           </>
         )}
-        <SettingsItem
-          icon={{ svgElement: HandshakeIcon, theme: 'default' }}
-          title="Rolle og rettigheter"
-          value="Daglig leder"
-          linkIcon
-        />
       </List>
     </Section>
   );
 };
 
-export const AccountToolbar = ({ id, type, isCurrentEndUser, favourite, onToggleFavourite }: AccountDetailsProps) => {
+export interface AccountToolbarProps extends AccountListItemProps {
+  id: string;
+  isCurrentEndUser: boolean;
+  favourite?: boolean;
+  onToggleFavourite?: (id: string) => void;
+}
+
+export const AccountToolbar = ({ id, isCurrentEndUser, favourite, onToggleFavourite }: AccountToolbarProps) => {
   return (
     <Flex spacing={2} size="xs">
       {!isCurrentEndUser && (
@@ -79,21 +139,6 @@ export const AccountToolbar = ({ id, type, isCurrentEndUser, favourite, onToggle
       <Button icon={InboxIcon} variant="outline">
         Gå til Innboks
       </Button>
-      {type !== 'group' && (
-        <Button icon={HandshakeIcon} variant="outline">
-          Tilgangsstyring
-        </Button>
-      )}
     </Flex>
   );
 };
-
-export interface AccountDetailsProps extends AccountListItemProps {
-  userId?: string;
-  alertEmailAddress?: string;
-  alertPhoneNumber?: string;
-  contactEmailAddress?: string;
-  contactPhoneNumber?: string;
-  address?: string;
-  items?: AccountListItemProps[];
-}